mejorando doc dupes
