--- monitoring/serializers.py
@@ -1,39 +1,0 @@
-from rest_framework import serializers
-
-from .models import *
-
-
-class CountrySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Country
-        fields = ['id', 'name']
-
-
-class OrganizationSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Organization
-        fields = ['id', 'name']
-
-
-class ProjectSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Project
-        fields = ['id', 'name']
-
-
-class ContactTypeSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ContactType
-        fields = ['id', 'name']
-
-
-class ContactSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Contact
-        fields = ['id', 'name']
-
-
-class EducationSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Education
-        fields = ['id', 'name']
